chore(cleanup): reorganize counterparty imports and refactor related modules
